requirements.txt
# Core dependencies for General Relativity tutorial
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
sympy>=1.12

# AI Assistant
anthropic>=0.39.0

# Interactive development
jupyter>=1.0.0
ipython>=8.12.0

# Animation for advanced visualizations
manim>=0.18.0

# For symbolic tensor calculations
einsteinpy>=0.4.0

# Environment management
python-dotenv>=1.0.0
